# packages/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CariCash Nova API
  version: 0.2.0
  description: |
    CariCash Nova — Barbados-based mobile payments platform API.

    ## Key Concepts
    - **Jurisdiction**: Barbados. Base currency: BBD. NO FX — cross-currency postings rejected.
    - **Closed-loop wallets**: Customers, agents, and merchants hold balances internally.
    - **Double-entry ledger**: Every transaction posts balanced journal entries (sum DR = sum CR).
    - **Idempotency**: All money-moving endpoints require `idempotency_key`. Retries with same key+payload return cached result; same key+different payload returns 409 DUPLICATE_IDEMPOTENCY_CONFLICT.
    - **Maker-checker**: Reversals, manual adjustments, overdraft facilities, fee/commission changes require dual approval (maker ≠ checker).
    - **Posting via Durable Object**: All money movement routes through a serialized Durable Object per wallet-per-currency.
  contact:
    name: CariCash Engineering
    email: engineering@caricash.bb
  license:
    name: MIT
servers:
  - url: /
    description: Current host
  - url: http://localhost:8787
    description: Local development (wrangler dev)
...
security: []
paths:
...
  /merchants:
    post:
      operationId: createMerchant
      tags: [Merchants]
      summary: Register a new merchant
      description: Create a merchant account with store_code, name, MSISDN, and PIN. Automatically creates a BBD wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantRequest'
            example:
              store_code: 'STORE002'
              name: 'Bob Shop'
              msisdn: '+12465557777'
              pin: '9012'
              correlation_id: '01HZ...'
      responses:
        '201':
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMerchantResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Merchant with this store code already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /merchants/{merchantId}:
    get:
      operationId: getMerchant
      tags: [Merchants]
      summary: Get merchant details
      description: Retrieve merchant profile by ID. Returns 501 (stub).
      parameters:
        - $ref: '#/components/parameters/merchantId'
      responses:
        '200':
          description: Merchant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorResponse'
        '501':
          description: Not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /wallets/{ownerType}/{ownerId}/{currency}/balance:
    get:
      operationId: getWalletBalance
      tags: [Wallets]
      summary: Get wallet balance
      description: |
        Retrieve current balance for a wallet. Balance is derived from the append-only ledger
        (sum of credits minus sum of debits on the account).
      parameters:
        - name: ownerType
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/OwnerType'
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
        - name: currency
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyCode'
      responses:
        '200':
          description: Balance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
              example:
                owner_type: CUSTOMER
                owner_id: '01HZ...'
                currency: BBD
                account_id: '01HZ...'
                balance: '1000.00'
                actual_balance: '1000.00'
                available_balance: '970.00'
                hold_amount: '30.00'
                pending_credits: '0.00'
                correlation_id: '01HZ...'
        '404':
          $ref: '#/components/responses/NotFound'
  /wallets/{ownerType}/{ownerId}/{currency}/all-balances:
    get:
      operationId: getAllWalletBalances
      tags: [Wallets]
      summary: Get all account balances for an owner
      description: |
        Retrieve balances across all account types (e.g. WALLET, CASH_FLOAT, COMMISSIONS_PAYABLE)
        for the specified owner and currency.
      parameters:
        - name: ownerType
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/OwnerType'
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
        - name: currency
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyCode'
      responses:
        '200':
          description: All balances response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllBalancesResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/customer/login
    AgentAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/agent/login
    MerchantAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/merchant/login
    StaffAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/staff/login (OAuth stub)
  parameters:
...
    merchantId:
      name: merchantId
      in: path
      required: true
      schema:
        type: string
      description: Merchant actor ID (ULID)
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Validation failed'
            code: VALIDATION_ERROR
            correlation_id: '01HZ...'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Not found'
            code: NOT_FOUND
            correlation_id: '01HZ...'
...
  schemas:
...
    OwnerType:
      type: string
      enum: [CUSTOMER, AGENT, MERCHANT, STAFF]
    ActorType:
      type: string
      enum: [CUSTOMER, AGENT, MERCHANT, STAFF]
...
    CurrencyCode:
      type: string
      enum: [BBD, USD]
      description: Supported currency codes. BBD is the base currency.
...
    KycStatus:
      type: string
      enum: [NOT_STARTED, PENDING, APPROVED, REJECTED]
    ActorState:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, CLOSED]
    AccountType:
      type: string
      enum: [WALLET, FEE_REVENUE, TAX_PAYABLE, COMMISSIONS_PAYABLE, OVERDRAFT_FACILITY, SUSPENSE, CASH_FLOAT]
    ErrorCodeEnum:
      type: string
      enum:
        - UNAUTHORIZED
        - FORBIDDEN
        - VALIDATION_ERROR
        - INSUFFICIENT_FUNDS
        - CROSS_CURRENCY_NOT_ALLOWED
        - LIMIT_EXCEEDED
        - KYC_REQUIRED
        - BLOCKLISTED
        - DUPLICATE_IDEMPOTENCY_CONFLICT
        - IDEMPOTENCY_IN_PROGRESS
        - MAKER_CHECKER_REQUIRED
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
        - UNBALANCED_JOURNAL
        - JOURNAL_NOT_FOUND
        - ACCOUNT_NOT_FOUND
        - MAKER_CHECKER_VIOLATION
        - APPROVAL_NOT_FOUND
        - APPROVAL_ALREADY_DECIDED
        - INVALID_CREDENTIALS
        - ACCOUNT_LOCKED
        - RATE_LIMIT_EXCEEDED
        - SESSION_EXPIRED
        - ACTOR_NOT_FOUND
        - ACTOR_SUSPENDED
        - DUPLICATE_IDENTIFIER
        - MISSING_REQUIRED_FIELD
        - RECONCILIATION_MISMATCH
        - INTEGRITY_VIOLATION
        - REPAIR_FAILED

    # ---- Error response ----
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCodeEnum'
        name:
          type: string
          description: Error class name
        correlation_id:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
          default: false

    # ---- Auth requests ----
...
    Actor:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActorType'
        state:
          $ref: '#/components/schemas/ActorState'
        msisdn:
          type: string
        agent_code:
          type: string
        store_code:
          type: string
        staff_code:
          type: string
        name:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
          format: email
        kyc_state:
          $ref: '#/components/schemas/KycStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ActorResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        correlation_id:
          type: string

    # ---- Customer ----
...
    CreateMerchantRequest:
      type: object
      required: [store_code, name, msisdn, pin]
      properties:
        store_code:
          type: string
        name:
          type: string
        msisdn:
          type: string
        pin:
          type: string
        correlation_id:
          type: string
    CreateMerchantResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        wallet_id:
          type: string
        correlation_id:
          type: string

    # ---- KYC ----
...
    BalanceResponse:
      type: object
      properties:
        owner_type:
          $ref: '#/components/schemas/OwnerType'
        owner_id:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        account_id:
          type: string
        balance:
          type: string
          description: Decimal string (e.g., "1000.00")
        actual_balance:
          type: string
          description: Materialized actual balance for the account
        available_balance:
          type: string
          description: Spendable or withdrawable balance
        hold_amount:
          type: string
          description: Amount currently on hold
        pending_credits:
          type: string
          description: Credits awaiting release
        correlation_id:
          type: string
    AccountBalanceEntry:
      type: object
      properties:
        account_id:
          type: string
        owner_type:
          $ref: '#/components/schemas/OwnerType'
        owner_id:
          type: string
        account_type:
          $ref: '#/components/schemas/AccountType'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        actual_balance:
          type: string
        available_balance:
          type: string
        hold_amount:
          type: string
        pending_credits:
          type: string
        updated_at:
          type: string
          format: date-time
    AllBalancesResponse:
      type: object
      properties:
        owner_type:
          $ref: '#/components/schemas/OwnerType'
        owner_id:
          type: string
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountBalanceEntry'
        correlation_id:
          type: string
...